# openapi.yaml
openapi: 3.0.0
paths:
  /api/user/profile:
    get:
      operationId: profile
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
      tags:
        - 👤 user
      security:
        - bearer: []
  /api/user/library:
    get:
      operationId: library
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLibrary"
      tags:
        - 👤 user
      security:
        - bearer: []
  "/api/user/start-reading/{id}":
    patch:
      operationId: startReading
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
      tags:
        - 👤 user
      security:
        - bearer: []
  "/api/user/finish-reading/{id}":
    patch:
      operationId: finishReading
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
      tags:
        - 👤 user
      security:
        - bearer: []
  "/api/user/toggle-save/{id}":
    patch:
      operationId: toggleSave
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: boolean
      tags:
        - 👤 user
      security:
        - bearer: []
  "/api/user/is-saved/{id}":
    get:
      operationId: isSaved
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: boolean
      tags:
        - 👤 user
      security:
        - bearer: []
  /api/user/admin/catalog:
    get:
      operationId: adminCatalog
      parameters:
        - name: searchTerm
          required: true
          in: query
          schema:
            type: string
        - name: page
          required: true
          in: query
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAdminCatalog"
      tags:
        - 👤 user
      security:
        - bearer: []
  "/api/user/admin/remove/{id}":
    delete:
      operationId: remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
      tags:
        - 👤 user
      security:
        - bearer: []
...
  "/api/book/by-id/{id}":
    get:
      operationId: infoById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfoByIdOutput"
      tags:
        - 📙 book
      security:
        - bearer: []
  "/api/book/ebook/{id}":
    get:
      operationId: ebookById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EbookByIdOutput"
      tags:
        - 📙 book
      security:
        - bearer: []
  "/api/book/admin/by-id/{id}":
    get:
      operationId: infoByIdAdmin
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInfoByIdOutput"
      tags:
        - 📙 book
      security:
        - bearer: []
  /api/book/admin/catalog:
    get:
      operationId: adminCatalog
      parameters:
        - name: searchTerm
          required: true
          in: query
          schema:
            type: string
        - name: page
          required: true
          in: query
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCatalogOutput"
      tags:
        - 📙 book
      security:
        - bearer: []
  /api/book/admin/create:
    post:
      operationId: create
      parameters: []
      requestBody:
        required: true
        description: Create book
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookDto"
      responses:
        "200":
          description: ""
      tags:
        - 📙 book
      security:
        - bearer: []
  "/api/book/admin/update/{id}":
    put:
      operationId: update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditBookDto"
      responses:
        "200":
          description: ""
      tags:
        - 📙 book
      security:
        - bearer: []
  "/api/book/admin/update-genre/{id}":
    put:
      operationId: updateGenre
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGenreDto"
      responses:
        "200":
          description: ""
      tags:
        - 📙 book
      security:
        - bearer: []
  "/api/book/admin/update-ebook/{id}":
    post:
      operationId: updateEbook
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/EBookType"
      responses:
        "200":
          description: ""
      tags:
        - 📙 book
      security:
        - bearer: []
  "/api/book/admin/remove/{id}":
    delete:
      operationId: remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
      tags:
        - 📙 book
      security:
        - bearer: []
  /api/auth/google-sign:
    post:
      operationId: googleSign
      parameters: []
      requestBody:
        required: true
        description: Sign in with google account
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignDto"
      responses:
        "200":
          description: Return access and refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
      tags:
        - 🔐 auth
  /api/auth/mail-register:
    post:
      operationId: register
      parameters: []
      requestBody:
        required: true
        description: Register new user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthDto"
      responses:
        "200":
          description: Return access and refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
      tags:
        - 🔐 auth
  /api/auth/mail-login:
    post:
      operationId: login
      parameters: []
      requestBody:
        required: true
        description: Login user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthDto"
      responses:
        "200":
          description: Return access and refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
      tags:
        - 🔐 auth
  /api/auth/refresh:
    post:
      operationId: refreshToken
      parameters: []
      requestBody:
        required: true
        description: Refresh access token
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshDto"
      responses:
        "200":
          description: Return access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
      tags:
        - 🔐 auth
  /api/storage/delete:
    post:
      operationId: delete
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilenameDto"
      responses:
        "200":
          description: File deleted
      tags:
        - 📁 storage
      security:
        - bearer: []
  "/api/storage/{folder}":
    post:
      operationId: upload
      parameters:
        - name: folder
          required: true
          in: path
          schema:
            enum:
              - ebooks
              - booksCovers
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadOutputDto"
      tags:
        - 📁 storage
      security:
        - bearer: []
  /api/parser/admin/catalog:
    get:
      operationId: adminCatalog
      parameters:
        - name: searchTerm
          required: false
          in: query
          example: The Hobbit
          schema:
            type: string
        - name: page
          required: false
          in: query
          example: 1
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookTemplateCatalogOutput"
      tags:
        - 📦 parser
      security:
        - bearer: []
  "/api/parser/admin/by-id/{id}":
    get:
      operationId: byId
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookTemplate"
      tags:
        - 📦 parser
      security:
        - bearer: []
  /api/parser/admin/parse:
    post:
      operationId: parse
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParserDto"
      responses:
        "201":
          description: ""
      tags:
        - 📦 parser
      security:
        - bearer: []
  /api/parser/admin/unfold:
    post:
      operationId: unfold
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Unfolded book content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UnfoldOutput"
      tags:
        - 📦 parser
      security:
        - bearer: []
  "/api/parser/admin/remove/{id}":
    delete:
      operationId: remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
      tags:
        - 📦 parser
      security:
        - bearer: []
...
info:
  title: Booknex
  description: ""
  version: "1.0"
  contact:
    name: Booknex
    url: https://github.com/kravchenko-anton/booknex-2-monorepo
    email: Github repository
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    UserProfile:
      type: object
      properties:
        id:
          type: number
          example: 1
        email:
          type: string
          example: email
        bookCount:
          type: number
          example: 1
        totalPageCount:
          type: number
          example: 1
      required:
        - id
        - email
        - bookCount
        - totalPageCount
    ShortBook:
      type: object
      properties:
        id:
          type: number
          example: 1
          description: book id
        title:
          type: string
          example: title
          description: book title
        picture:
          type: string
          example: picture
          description: book picture
        author:
          type: string
          example: author
          description: book author
      required:
        - id
        - title
        - picture
        - author
    UserLibrary:
      type: object
      properties:
        readingBooks:
          type: array
          items:
            $ref: "#/components/schemas/ShortBook"
        finishedBooks:
          type: array
          items:
            $ref: "#/components/schemas/ShortBook"
        savedBooks:
          type: array
          items:
            $ref: "#/components/schemas/ShortBook"
      required:
        - readingBooks
        - finishedBooks
        - savedBooks
    shortGenre:
      type: object
      properties:
        id:
          type: number
          example: 1
          description: genre id
        name:
          type: string
          example: name
          description: genre name
      required:
        - id
        - name
...
    UserAdminCatalog:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CatalogUserOutput"
        canLoadMore:
          type: boolean
          example: true
        totalPages:
          type: number
          example: 1
      required:
        - data
        - canLoadMore
        - totalPages
...
    InfoByIdOutput:
      type: object
      properties:
        id:
          type: number
          example: 1
          description: book id
        title:
          type: string
          example: title
          description: book title
        picture:
          type: string
          example: picture
          description: book picture
        author:
          type: string
          example: author
          description: book author
        description:
          type: string
          example: description
          description: book description
        genres:
          type: array
          items:
            $ref: "#/components/schemas/shortGenre"
      required:
        - id
        - title
        - picture
        - author
        - description
        - genres
...
    EbookByIdOutput:
      type: object
      properties:
        picture:
          type: string
          example: image.png
          description: book cover
        title:
          type: string
          example: title
          description: book title
        file:
          description: book file
          type: array
          items:
            type: string
        chapters:
          description: book chapters
          type: array
          items:
            $ref: "#/components/schemas/Chapter"
      required:
        - picture
        - title
        - file
        - chapters
...
    AdminInfoByIdOutput:
      type: object
      properties:
        id:
          type: number
          example: 1
          description: book id
        title:
          type: string
          example: title
          description: book title
        picture:
          type: string
          example: picture
          description: book picture
        author:
          type: string
          example: author
          description: book author
        createdAt:
          type: string
          example: 2021-07-01
          description: book created at
        updatedAt:
          type: string
          example: 2021-07-01
          description: book updated at
        visible:
          type: boolean
          example: true
          description: book visibility
        ebook:
          type: string
          example: ebook
          description: book ebook
        description:
          type: string
          example: description
          description: book description
        popularity:
          type: number
          example: 100
          description: book popularity
        pages:
          type: number
          example: 100
          description: book pages
        activities:
          description: book activities
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        _count:
          description: book count
          allOf:
            - $ref: "#/components/schemas/CountOutput"
        review:
          description: book review
          type: array
          items:
            $ref: "#/components/schemas/Review"
        genres:
          type: array
          items:
            $ref: "#/components/schemas/shortGenre"
      required:
        - id
        - title
        - picture
        - author
        - createdAt
        - updatedAt
        - visible
        - ebook
        - description
        - popularity
        - pages
        - activities
        - _count
        - review
        - genres
...
    AdminCatalogOutput:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        canLoadMore:
          type: boolean
          example: true
          description: can load more
        totalPages:
          type: number
          example: 1
          description: total pages
      required:
        - data
        - canLoadMore
        - totalPages
    EbookChapter:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        text:
          type: string
      required:
        - id
        - name
        - text
    EBookType:
      type: object
      properties:
        title:
          type: string
        id:
          type: number
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/EbookChapter"
      required:
        - title
        - id
        - chapters
    CreateBookDto:
      type: object
      properties:
        title:
          type: string
          description: Title of the book
          example: The Great Gatsby
        author:
          type: string
          description: Author of the book
          example: F. Scott Fitzgerald
        description:
          type: string
          description: Description of the book
          example: The Great Gatsby is a novel by the American author F. Scott Fitzgerald.
            First published in 1925, it is set on Long Island's North Shore and
            in New York City from spring to autumn of 1922.
        ebook:
          type: array
          items:
            $ref: "#/components/schemas/EBookType"
        pages:
          type: number
          description: Number of pages in the book
          example: 300
        popularity:
          type: number
          description: Number of goodRead reviews
          example: 1000000
        genres:
          description: Array of genres
          example:
            - 1
            - 2
            - 3
          type: array
          items:
            type: number
        picture:
          type: string
      required:
        - title
        - author
        - description
        - ebook
        - pages
        - popularity
        - genres
        - picture
    EditBookDto:
      type: object
      properties:
        picture:
          type: string
        title:
          type: string
          description: Title of the book
          example: The Great Gatsby
        author:
          type: string
          description: Author of the book
          example: F. Scott Fitzgerald
        description:
          type: string
          description: Uploaded picture
          example: picture.jpg
        pages:
          type: number
          description: Number of pages in the book
          example: 300
        visible:
          type: boolean
          description: Is book visible
          example: true
        popularity:
          type: number
          description: Number of goodRead reviews
          example: 1000000
    UpdateGenreDto:
      type: object
      properties:
        genres:
          description: Array of genres
          example:
            - 1
            - 2
            - 3
          type: array
          items:
            type: number
    SignDto:
      type: object
      properties:
        socialId:
          type: string
          description: Social id
          example: "1234567890"
      required:
        - socialId
...
    AuthResponseDto:
      type: object
      properties:
        accessToken:
          type: string
          description: Access token
          example: "1234567890"
        refreshToken:
          type: string
          description: Refresh token
          example: "1234567890"
        type:
          type: string
          description: type of auth
          example: login
        user:
          description: User data
          allOf:
            - $ref: "#/components/schemas/AuthUserDto"
      required:
        - accessToken
        - refreshToken
        - type
        - user
    AuthDto:
      type: object
      properties:
        email:
          type: string
          description: User email
          example: test@gmail.com
        password:
          type: string
          description: User password
          example: password
      required:
        - email
        - password
    RefreshDto:
      type: object
      properties:
        refreshToken:
          type: string
          description: Refresh token
          example: "1234567890"
      required:
        - refreshToken
    FilenameDto:
      type: object
      properties:
        filename:
          type: string
      required:
        - filename
    UploadOutputDto:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    BookTemplate:
      type: object
      properties:
        id:
          type: number
          description: id of the book
        title:
          type: string
          description: title of the book
        author:
          type: string
          description: author of the book
        description:
          type: string
          description: description of the book
        picture:
          type: string
          description: picture of the book
        pages:
          type: number
          description: pages of the book
        popularity:
          type: number
          description: popularity of the book
        genres:
          type: array
          items:
            $ref: "#/components/schemas/shortGenre"
      required:
        - id
        - title
        - author
        - description
        - picture
        - pages
        - popularity
        - genres
    BookTemplateCatalogOutput:
      type: object
      properties:
        data:
          description: book template
          type: array
          items:
            $ref: "#/components/schemas/BookTemplate"
        canLoadMore:
          type: boolean
          description: can load more
        totalPages:
          type: number
          description: total pages
      required:
        - data
        - canLoadMore
        - totalPages
    ParserDto:
      type: object
      properties:
        url:
          type: string
          description: url of the parser
        page:
          type: number
          description: page of the parser
      required:
        - url
        - page
    UnfoldOutput:
      type: object
      properties:
        id:
          type: number
          description: id of the chapter
        name:
          type: string
          description: name of the chapter
        text:
          type: string
          description: text of the chapter
      required:
        - id
        - name
        - text
...